Add script to test firmware
Add the script test_all.py to test the firmware of all connected
daplink boards.  In specific this script tests the following:
1. Mass storage is functioning
2. Serial port is working
3. HID is working

This patch is the first pass of automated dap testing.  In
the future the tests will be expanded for all endpoints.  TODOs have
been left to indicate desired future changes.

Note:
-The file basic_test.py was taking from pyOCD and is used as a
starting point for HID testing.  In the future this test will be
heavily modified.
-The file mbedapi.py was taken from
https://developer.mbed.org/teams/mbed/code/mbed-API-helper/
and modified to to be more usable as a script.  It was also
modified to more closely match pep8.