Add 'Version' (V) option to the URL parameters
version.c allows insertion of parameters to a URL, however some partners
may not want any personally identifyable information sent to a web service.

This commit adds a 'V' format that adds just the firmware version and device
type but not the MAC or any other specific information

At this time, using Xs instead of a shorter string is done to preserve the
length of the string.