Fix interface erase bug in maintenance mode
Loading an application hex file in maintenance mode causes the
interface to get erased.  This is because the interface is erased as
soon as a valid hex is detected, regardless of whether the starting
address is correct.

This patch fixes the problem by delaying the flash erase to just before
the first page is programmed.  This allows the address to be checked
before an erase is performed.