Fix bug showing error when IF is programmed
When programming a new interface hex file in maintenance mode an error
is displayed and the device stays in maintenance mode.  This is because
the interface has config data which lies outside of the valid address
range.  This causes programming to indicate that an error occurred,
even though the interface was loaded.

This patch fixes the problem by declaring the interface config section
as 'UNINIT', along with making the config variable 'zero_init'.  The
attribute 'zero_init' is required to prevent config from showing up
in the hex file.  The scatter attribute 'UNINIT' is required to prevent
__scatterload_zeroinit from trying to initialize config on boot, which
causes a hard fault.