Prevent reset line from being held low in BL
Keep the target unpowered in the K20DX boot loader to prevent it from
trying to drive the reset line low.  This fixes a problem on the K22F
and K64F freedom boards where after loading new interface firmware
the device stays in bootloader mode because the reset line is held
low by the target.