Fix trigraph warnings
The sequence '??=' is interpreted as a trigraph.  This instructs the
compiler to replace '??='  with the character '#' before compiling
the code.  The file system_MK20D5.c uses this sequence in the code's
comments, without the intent of being interpreted as a trigraph.
Keil warns about this since trigraphs are disabled by default in
gcc mode.

This patch replaces all instances of the sequence '??=' with '?=' to
prevent them from being interpreted as a trigraph and printing a warning.