Remove extra sectors
Reduce the number of sectors in the fat table from 8 to 6.
This is the correct number of sectors when 'Fail.txt' is being
displayed, which allows windows chkdsk to pass when in this state.
Running chkdsk when the fail file is not present will still result
in warnings about lost cluster chains.