Revert reset config filename
Revert the reset config filename back to a version which does not
start with a period.  This allows the file to be treated properly on
Windows and Mac devices.

Note:  Having a period in the filename appears to be valid and passed
the following filesystem verification tools - chkdsk (Windows),
fsck (Linux) and volumeVerify (Mac).  Even though these programs
indicated that there were no problems with the filesystem, both
Windows and Mac devices reported that the file did not exist and
could not be deleted.  On Linux this file could be deleted without any
problems.