Fix reset problems on K20DX interface hardware
Move the code which forwards the reset pin into the weakly referenced
function target_forward_reset().  On the K20DX overwrite this function
with one which does nothing, since the reset button is already tied to
the target reset pin.

This fixes programming problems on the K20DX when the target
processor has invalid code causing the reset line to oscillate.  This
is because the reprogramming sequence is not interrupted by a
reset triggered by the state machine in main.c.

[squash] fix reset problems