Switch between BL and IF by deleting file
Add the ability to switch between bootloader and interface by deleting
a file on the filesystem.  This enables new interface code to be loaded
from automated test scripts.  Rebooting into bootloader mode is done
through a shared RAM region that the interface uses to signal to
the bootloader to stay in bootloader mode.

This patch also simplifies the configuration settings by making them
a packed struct and using sizeof directly, rather than obtaining
the size by getting the position of the last member.  This change
retains compatibility with the previous settings format, but makes
the code easier to understand.

Finally, this patch adds a delay to the K20DX family of boards
when booting.  This allows the target voltage rails to fully drain
if they had been previously powered, preventing reset from being
driven low by the target.