Overhaul mass storage and fix load with shutils
When using shutils to load a large firmware image on to a board,
the transfer is flushed partway through.  To the device this appears
as if the current file write is finished.  Because of this daplink
terminates the transfer and reconnects USB.  This causes the
target firmware to only get partially loaded, corrupting it.

This patch fixes the problem by waiting for a short period of time
after data is no longer sent over USB before disconnecting.  If more
data arrives then the programming resumes.

This patch updates the handling of mass storage data to treat all
cases the same.  Regardless of programming success or failure,
disconnect occcurs the same way.  Once a termination event
occurs, such as the end of a hex file arriving or a flash
error, mass storage packets are no longer processed.  The device
will wait until the PC stops sending data, and then a fixed amount
of time after that USB is disconnected and reconnected.

This change consists of the following modifications:
-Combine flash writes so writes only occur in one location
-Remove use of forced USB disconnect to prevent transfers from
    being aborted midway through
-Add timeout so if a binary or hex file is not fully sent the drive
    will disconnect and show fail.txt
-Move file_transfer_state and its type into usbd_user_msc.c to limit
    its scope
-Add DELAY event to signal that the USB disconnect should be delayed
-Add comments by flash fail contents and a compiler assert to ensure
    it stays in sync with the enumeration above it