Fix hang when loading a binary smaller than 4K
On Windows when a binary file smaller than 4K is loaded the file
entry is written before the file contents.  When the file contents
arrive the variable file_transfer_state.amt_to_write gets overwritten
and set to 0xffffffff in the transfer initialization code.  This causes
the board to think that there is more data coming, even after all
the file contents have been written.  Because of this mass storage
is never detached and re-connected.

This patch fixes the USB hanging by only setting
file_transfer_state.amt_to_write to 0xffffffff if it has not already
been set.