Fix USB ID reported when in bootloader mode
When in bootloader mode the USB ID reported is invalid.  This is
because the ID is not getting properly initialized.  This patch
adds the initialization to the startup sequence in main.c.

This bug was introduced in the following patch:
"b98a2520ef4083db517fd2e669503699ef4e0ee1" -
"Sync src with 7907f65e48c34f55a5362bed51d67d129b1442eb
from /sg-/cmsis-dap and flatten some common directories"