Update vect table to identify build, ver and HDK
Add DAPLINK_BUILD_KEY to identify if the current image is a bootloader
or an interface.  Add DAPLINK_HDK_ID to identify which kind of
firmware is compatible and can be safely loaded on a device.
Add compiler asserts to check the HDK ID in files with HDK
specific information.  Finally, add DAPLINK_VERSION to indicate
the build version.