Abstract filesystem interface from type
Remove all references to fat from virtual_fs.h.  In particular remove
the FatDirectoryEntry_t.  To obtain information about a file access
methods should be used on that file's handle.

Also fix a bug in the initialization for default_transfer_state by
filling in every field with the appropriate value.