Prevent USB enumeration failure in the future
At the time of this patch USB is working correctly.  This patch is
to prevent enumeration problems from appearing in the future.
It was found that CPU intensive operations, such as running a CRC
on a thread of higher priority than the endpoint 0 thread causes
enumeration to fail.  This patch bumps the priority of endpoint 0,
so even under heavy CPU load enumeration occurs correctly.