Fix crash when programming settings
Round the size to flash up to a page boundary so programming regions
smaller than a page work correctly.  Also return failure if a size of
zero is passed in.  Prior to this a size smaller than a page would
cause NumPagesLeft to be shifted to zero.  This resulted
in a overflow in the while loop below.