Increase speed CMSIS-DAP speed of atsam3u
Change the High speed hid report interval from 6 (x125us)
to 1 (x125us).  This increases the maximum theoretical
throughput to 8x that of full speed USB.  Also bump the priority
of the USB threads.  This allows hid packet processing to be
pre-empted by the USB thread reading hid packets.  This allows
the maximum number of packets to be properly queued.