Run USB on single thread to avoid threading bugs
Turn off the RTX mode of the USB driver and use it in the context of a
single thread.  Since all USB processing is done on a single thread
this fixes synchronization problems causing trouble for high speed
USB.

This patch also has the added benefit of greatly reducing RAM usage
since there is no longer the need for multiple USB threads.