Add runtime asserts
Add support for runtime asserts.  Asserts can be used in the bootloader
or interface by using util_assert(expression).  The assert will appear
on the filesystem on the next remount.  The assert will reside there
until it is explicitly deleted, or power is removed.

Also add action file "REFRESH.ACT" to cause an update of the filesystem
and action file "ASSERT.ACT" to trigger an assert for testing purposes.