Optimize uart_set_configuration for speed
Optimize the call to uart_set_configuration in the lpc11u35 for speed
by calculating an approximate value.  While this prevents the most accurate
possible configuration from being selected, the baudrate
selected is accurately enough to function correctly.

This change is made to fix long uart configuration times which cause
some mbed-driver green tea tests to fail.  Two tests in particular
that fail because of this are 'serial_interrupt' and 'echo'.