Remove use of __inline for C99 compatibility
Keil 5.16 changed how __inline is implemented in C99 to more closely
match that of the C specification.  The introduced undefined function
errors in usbd_core.c.  This patch fixes that by replacing the use of
"__inline" with "static inline".

More info on this can be found here:
http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.faqs/ka15831.html

Taken from the above link:

$ armcc test.c -o test.axf -O0 --c99
  Error: L6218E: Undefined symbol inc (referred from test.o).
This error is expected and caused by the special C99 specification for
the inline keyword. When building for C90 (using __inline keyword
instead of inline) or C++, no error appears. In C99, a function that
is marked with inline is an inline function.  An inline function in C99
has an inline definition. An external definition (out-of-line copy of
the function definition) is never generated. Whereas in C90 and C++
an external definition of the function is generated.

To use inline in C99 properly, it is recommended to specify the
function with static inline, or put the definition of the function info
a header file and specify it with extern inline.