Fix assert from loading target image in BL
When in bootloader mode and the flash type is thought to be type target
flash an error is triggered.  This is because there is no target
programming support in bootloader.  Since the image being loaded could
be one from a 3rd party vendor that is meant to run as an application,
treat this as if it were an image to be loaded into internal flash.