Fix bug with end of transfer detection
Set the flag indicating that the transfer can be finished if either the
stream can be closed or the stream is finished.  This fixes file
transfer timeout errors on Mac computers found by automated testing.

Also keep track of the size transferred and size processed by the
stream.  This allows the error messages shown to more closely reflect
what actually happened.