Refactor virtual FS user into manager and user
Split virtual_fs_user into vfs_user and vfs_manager.  The code in
vfs_manager is complex, easy to break if not careful, and should
rarely need to be updated.  It is also generic, and is not DAPLink
specific.  The file vfs_user on the other hand is DAPLink specific.
It can be changed to easily reconfigure the filesystem, and carries
very little risk of introducing new bugs.