Remove nonvolatile settings from bootloader
Don't allow the bootloader to read or write any configurable settings
stored in rom.  This greatly increases the safety of the bootloader
since it no longer has state which persists across power cycles.

With this patch all settings in rom are written and read by the
interface alone.  The bootloader has stubbed calls for these settings
which return the same fixed value.