Fix intermittent failures when unlocking Kinetis
When unlocking Kinetis devices the function target_set_state is made to
set the mode to RESET_HOLD.  This leads to a call to swd_init which in
turn leads to a call to DAP_Setup which reconfigures all SWD pins to
be in a high-z state.  This causes the pins to float until
PORT_SWD_SETUP has a chance to re-initialize these pins.  On some
boards this causes a glich which is interpreted as a clock pulse.
This causes the next SWD transaction to fail.

This patch fixes that problem by replacing the call to
target_set_state(RESET_HOLD) with swd_set_target_reset(1) which does
not re-initialize IO.

Finally, this patch leaves a note inside swd_init for further
investigation.