Fix hex loading error
If the EOF record of a hex file is on its own sector with no data then
ERROR_SUCCESS_DONE will never get returned, this causes the code
to think the transfer is still in progress and give the error
"The transfer timed out.".  This patch fixes that bug by returning
ERROR_SUCCESS_DONE even if no bytes were written.