Fix crash during DAPLink specific testing
One of the DAPLink specific tests verifies that a CRC is not required
when loading an interface image.  It does this by intentionally
corrupting the image and then verifying that the image was loaded
successfully.  This is problematic because in some cases the interface
image loaded does not run.  This causes boards to get into a bad
state when running this test.

To fix that problem this patch skips that test when loading interface
binaries.  An earlier tests still checks that interfaces can be loaded
with a bad CRC by corrupting the CRC at the end of the image rather
than data in the image.