Fix serial_test.py by making write asynchronous
Under certain configurations serial_test.py fails.  This is likely
because the receive serial buffer on the OS fills before the write
completes and the test is able to read data out of the buffer.
This patch fixes that by making the write asynchronous so the read is
performed at the same time.