Reset target on boot only as necessary
Remove the call to target_set_state(RESET_RUN) from the common boot
sequence.  Call this function in the implementations of
prerun_target_config() which halt the target.  This allows targets
with the default implementation of prerun_target_config() to avoid
an extra reset on boot.

Also add a call to DAP_Setup() in the init sequence so all SWD GPIO
and state is properly initialized.  This was being done implicitly
by target_set_state(RESET_RUN).