Always use one bank per USB endpoint on sam3u
The sam3u USB driver assigns the number of buffer banks an endpoint has
based on endpoint number.  This patch removes that logic and instead
sets all endpoints to use only one bank.  This both allows for larger
endpoints and allows endpoints to be reordered in usb_config without
the risk of breaking due to a different number of banks.