Set all bulk EP on sam3u to 512 to match USB spec
High speed bulk endpoints are required to have an endpoint size of
512 bytes.  This patch updates the bulk CDC endpoints of the sam3u
to this size.

5.8.3 of the USB 2.0 specification:
"All Host Controllers are required to have support for 8-, 16-, 32-,
and 64-byte maximum packet sizes for full-speed bulk endpoints and
512 bytes for high-speed bulk endpoints. No Host Controller is required
to support larger or smaller maximum packet sizes."

This patch fixes problems seen when running serial test on OSX 10.10.5.