Reset MSC endpoint on USB reset to fix bug
If a USB reset occurs in the middle of a SCSI command
mass storage can get in a state where it thinks the transfer
is still ongoing.  This cases new SCSI commands to get ignored
by the device.  This causes the PC to disable the MSC endpoint.

This patch fixes that problem by properly reseting the MSC endpoint
when a USB reset occurs.