Fix bug with handling SWD protocol errors
When an SWD protocol error occurs the SWDIO line is unintentionally
left in input mode. This causes further transactions to fail.

This patch corrects this by setting SWDIO to output after a protocol
error. This also aligns with the more recent versions of CMSIS-DAP
(CMSIS-SP-00300-r4p5-00rel0) which contain this change.