Fix uart assert on k20dx and kl26z
When chaning settings in the middle of a transfer a uart assert gets
triggered. This occurs because the TX interrupt is left unmasked when
there is no data to send after the TX buffers have been cleared due
to a settings change.

This patch explicitly disabled the TX interrupt when settings have
changed to ensure this does not happen. It also asserts that the TX
interrupt is disabled when the uart buffers are cleared.