Add serial settings change smoke test
Add a smoke test for changing baudrate in the middle of a transfer.
Move the block transfer test to the last test, so if any of the
previous smoke tests put the device in a bad state, this will catch
it. Finally, trigger a DAPLink refresh to check for asserts at the end
of the serial test.