Don't remount if file is deleted before data sent
When a file with a '.bin' or '.hex' extension is found then the remount
timer starts. Abort the remount if the file is deleted before any data
is sent.

When copying files with attributes that do not exist on FAT16 windows
goes through the following steps:
1. Create an uninitialized file of the correct size
2. Delete the file created in step 1
3. Create new file with size 0
4. Prompt user:
    "Are you sure you want to copy this file without its properties"
5. If yes then send file data, if no the delete file

This causes problems because after the file in step 1 is deleted
DAPLink detects this as a transfer error and remounts. By aborting
the transfer instead of triggering an error files with incompatible
attributes can be properly copied.

This problem can be reproduced by saving a target binary to Dropbox's
sync folder and then from there copying it to DAPLink.