lpc11u35 - Prevent dropped USB SETUP packets
When a SETUP packet comes immediately after a control transfer with a
data in stage there is a chance that the OUT packet sent by the host
at the end of the transfer will be dropped. This is because there is no
way to tell that the SETUP and OUT are both pending so only the SETUP
packet gets processed.

This patch records whether an OUT packet is expected or not on the
control endpoint. If it is then it ensures that the OUT packet is
processed first.