lpc11u35 - Fix lockup in UART isr
If a call to uart_set_configuration() occurs at the right time when
receiving data then the code will get stuck in a loop trying to
handle receiving a character. This is because the DLAB bit in the
LCR register is set, which prevents the receive bit in the LSR
from getting cleared.

This problem arises because interrupts are unintentionally
enabled inside the critical section of uart_set_configuration()
by the call to uart_reset().

This patch fixes this problem creating and using a static reset
function which does not disable interrupts. This is then called
in uart_set_configuration() so interrupts stay disabled as expected.