lpc11u35 - fix entering ISP mode from application
When entering ISP mode from a running application the main stack
(MSP) is set to the correct value, but the processor is still running
using the process stack (PSP). This causes a crash in ISP mode.
This patch fixes this by switching the current stack to the MSP before
jumping to ISP mode.