Stay in bootloader mode if firmware crashes
Replace the previous hold in bootloader function with a set reset
mode function. There are three supported modes - bootloader, interface
and error. The error mode is the same as bootloader mode, but is used
to indicate that an error occurred.

On boot the next mode is set to the error mode.  If an unexpected crash
or reset occurs the device will enter error mode. otherwise, if the
device switches modes normally using util_reset_to_mode() a mode
switch will occur as expected.

This makes automated device recovery much easier. If bad interface
firmware causes a crash the device will automatically enter a mode
in which recovery can be done.