Fix incorrect mass storage overflow assert
In USBD_MSC_MemoryWrite if Offset and BulkLen are both 0 then then
there is an integer underflow and an assert is falsely triggered. This
patch corrects that logic so the assert is is only triggered if there
would have been a buffer overflow.

Prior to this patch this assert would occur during a write if the media
state changes to not ready after the CBW has been read but before
any data has been transferred.