Prevent lockup when target DP won't powerup
If for some reason a targets DP won't powerup, DAPLink will get stuck
in a loop waiting for this to occur. Add a timeout to prevent this
timeout and to allow DAPLink to boot normally.

This problem can be reproduced on the FRDM-KW24D if target firmware
sets pin PTB19 low immediately after boot.