atsam3u - fix sporadic USB reset during MSD xfer
Update the cached value of the endpoint status register (EPSTA) when an
endpoint is enabled/disabled, stalled/unstalled or reset. This fixes a
bug causing the device to sometimes send invalid packets, leading to a
USB reset and a re-enumeration.

This problem started occurring with the patch:
"atsam3u - Prevent dropped USB SETUP packets" -
2bb5379717dcf67119165734f35a8a193cc35d51