Prevent stuck serial from DAPLink to PC
If serial data is sent from the target to DAPLink before enumeration
completes the code will write this data to an endpoint, even though
the endpoint is not enabled. This causes DAPLink to enter a state where
it will receive serial data from the PC but will not send serial data to
the PC.

This patch prevents data from being written to an uninitialized
endpoint by preventing access to the endpoints before USB has
been configured. This prevents serial from getting stuck.