Update board info when retrying a test
If an IOError causes a test to retry the board info must be updated.
This is because a remount might have been missed, causing the
cached remount count and board mode to be out of sync which
causes a test failure. This patch prevents that from occurring by
updating the board info if this exception occurs.