Disable interrupts when performing flash ops
Update the atsam3u and kinetis flash drivers so they disable
interrupts when erasing or programming flash. This prevents
a crash from occurring if an interrupt occurs during flashing
when the flash cannot be read.

This also fixes a watchdog reset on K20 interfaces which occurs when
an interrupt interrupts the watchdog unlock sequence in the flash
"Init" function.  The addition of this critical section prevents the
watchdog operation from being interrupted.