Update Kinetis code to init debug without reset
Remove the code which asserts reset when initializing debug.
This workaround is no longer necessary as programming failures due to
the processor entering a secure state on the first reset after
programming are fixed with the change
"Reset and halt target after programming".
With this change, any target state changes due to new flash contents
happen as part of the programming sequence rather than on the first
reset after programming.

How to reproduce the failure this was intended to fix:
1. Checkout this commit
2. Revet the change "Reset and halt target after programming"
3. If it isn't 0 turn DAPLink's "Auto Reset" feature off
4. Build firmware for a Kinetis board and truncate it to 0x400 bytes
  so it has a valid vector table but does not contain the security bits
5. Load this binary image. This initial load should be successful
6. Load the same image again without letting the target run. After
  programming FAIL.TXT will be displayed with the error message
  "The interface firmware FAILED to reset/halt the target MCU". If you
  see this then have successfully reproduced this issue.