IO_Config.h: Modify the way it is included
This header file defines gpio and pin,
but is included in many places that are not needed.
Because these places need to introduce cmsis_core,
and because cmsis_core depends on the IRQn_Type definition of soc,
the cmsis_core header file cannot be directly imported in the corresponding place.
So modify the response of the .yaml file with the --preinclude parameter
to globally include the definition of the soc containing the cmsis_core header file.
"#include "IO_Config.h"" is safely removed.
Added only where gpio and pin definitions are needed

Signed-off-by: Findlay Feng <i@fengch.me>