daplink: Rename weak definition file
When using static __inline to define a function in armcc,
a file-related symbol is generated.
If it is the same file name,
it is possible to generate duplicate definition errors.
Weakly defined files have a very high probability of being replaced by files of the same file name,
resulting in symbol conflicts during the linking process.

Signed-off-by: Findlay Feng <i@fengch.me>