LPC55xx: override validate_bin_nvic for bootloader.
- Because the LPC5xx series raises bus faults on attempts to read from
erased sectors in internal flash, the standard validate_bin_nvic() has
to be overridden to prevent the bootloader from faulting when it tries
to check if there is a valid interface app in flash.
- This is done by supplying a target family and custom NVIC validator
function.
- The standard NVIC validation code is split out from validate_bin_nvic()
into validate_bin_nvic_base() so it can be called from the custom
validator once it is determined that the sector is programmed, or if
the given address is not in internal flash.
- Provide an overridden implementation of flash_is_page_readable() that
uses the flash blank check command to see if the page is erased.