Fix dropped uart TX data on nRF52840
If uart data is being transmitted and uart_write_data is called
it will start a new transfer and clobber the currently transmitting
data.

This patch prevents that by only starting a transfer if there is not
one occurring already. Additionally to prevent race conditions between
interrupt and non-interrupt code, this patch only starts TX from the
uart interrupt handler.