LPC55xx: Fix ARM_USART_ABORT_{SEND,RECEIVE} control calls
The FIFOINTENSET register is used to read the currently enabled
interrupts in the USART block and to enable new interrupts by writing a
1, but it can't be used to disable them. For that we need to use the
FIFOINTENTCLR register.

USART_TransferAbortSend and USART_TransferAbortReceive are the
equivalent functions using the same API that the rest of this driver is
using so it makes more sense to use those functions instead.