LPC55xx: Fix uart_write_data() with pending transfers.
`uart_write_data()` uses a circular buffer to enqueue next TX bytes and
a single `cb_buf.tx` byte for the "current" byte it is sending. This
function was always calling `USART_INSTANCE.Send` without checking
whether a transfer was already in progress, which would cause the
fsl_uart_cmsis.c to override the current transfer if there was one in
progress. This would cause the next byte in the circular buffer (the
one returned by `pop`) to be discarded when `uart_write_data()` is
called and there is already data pending in the circular buffer. This
bug can be reliably reproduced when issuing multiple writes on the
/dev/ttyACM0 device on a short period of time.

This patch fixes the problem by adding a flag that tells whether the
current transfer is still ongoing or not. At the moment, since all the
`Send()` calls use a size of 1 this flag can only be 0 or 1, but it will
increased in the future for a more performant implementation.

Tested on Linux sending multiple writes of various lengths and checking
the receiving end on a different USB-to-UART adapter.