LPC55xx: Improve UART TX performance.
Without this patch, transmitting at 921600 baudrate with 8n1 mode yields
to about 49.8KiB/s instead of the theoretical limit of 90 KiB/s. The
reason for this is because between the rising edge of the last data bit
(start of the STOP bit) and the falling edge before the START bit of the
next transfer we observe about 9.7us gap while the minimum is just the
length of a single STOP bit (about 1.1us). This gap is attributed to the
CPU time running the event handler (uart_handler) and scheduling a new
transfer. The performance was measured using a USB-to-UART probe running
with: `picocom -b 921600 /dev/ttyUSB0 | pv >/dev/null` and also looking
at the TX signal on the oscilloscope.

This commit creates two new functions in the circular buffer module that
let you peek into the internal buffer (avoiding a memcpy) and more
importantly allowing to return larger buffers to be passed to the UART
driver in a single Send() call. The UART module in the chip already has
a small FIFO that would allow the non-blocking transfer to enqueue the
buffer with 0 delay between consecutive bytes of the same transfer.
Given the 512 byte circular buffer size and the maximum 128 byte
transfer size this means that only one every 128 bytes would have a
~9us bigger gap with the previous one than needed. At a baudrate of
921600, 128 bytes take 1389 us to transmit so a 9 us delay is about a
1% performance degradation.

After this patch the effective TX bandwidth goes up 89.4 KiB/s, just shy
of the theoretical maximum of 90 KiB/s and consistent with the expected
performance.