lpc55xx_hic: Abort ongoing TX when changing the UART config
Changing the UART config (baudrate, parity, etc) was clearing the RX and
TX buffers, but if there was an ongoing TX transfer it would try to
remove the bytes from the TX buffer after the transfer completes. This
patch cancels any ongoing TX transfer when clearing the buffers on
config changes and UART reset.