lpc55xx_hic: Fix circular buffer and re-apply TX performance improvements (#932)
* Re-apply "LPC55xx: Improve UART TX performance."

This reverts commit 08d6cb507af5e3bcdc1bf90fe338449dd039f85d.

* lpc55xx_hic: Fix circ_buf asserts

The assert should be checking that there are more or the same number of
bytes in the circular buffer than those that are being removed, not the
other way around. This was causing an erroneous assert condition in the
UART interrupt handler.

* lpc55xx_hic: Abort ongoing TX when changing the UART config

Changing the UART config (baudrate, parity, etc) was clearing the RX and
TX buffers, but if there was an ongoing TX transfer it would try to
remove the bytes from the TX buffer after the transfer completes. This
patch cancels any ongoing TX transfer when clearing the buffers on
config changes and UART reset.