microbitv2: Replace MAIN_USER_EVENT state with a on-demand function.
So rather than go through a specific main_shutdown_state state to set
the "user event" (raising an interrupt and sending I2C data),
it just does that action directly when the button is long-pressed
or when the wake_from_usb flag is checked.

Also sets the interface_power_mode default state to 'Running'.
Not MB_POWER_DOWN, as until a power mode is requested the device
is awake.